+++ src/routes/mod.rs
@@ -2,6 +2,7 @@ pub mod container_create;
 pub mod container_remove;
 pub mod container_stop;
 pub mod containers_list;
+pub mod events_ws;
 pub mod exec;
 pub mod metrics;
 pub mod read_file;
